refactor(vault_client): split monolithic file into separate modules
Extracted 6 classes from vault_client.py (890 lines) into dedicated files:
- VaultKv2Secrets -> vault_kv2_secrets.py (114 lines)
- VaultKv1Secrets -> vault_kv1_secrets.py (84 lines)
- VaultPki -> vault_pki.py (212 lines)
- VaultAclPolicies -> vault_acl_policies.py (99 lines)
- VaultNamespaces -> vault_namespaces.py (203 lines)
- Secrets -> vault_secrets.py (29 lines)

vault_client.py now contains only VaultClient class and re-exports all
classes for backward compatibility (245 lines).

All existing imports continue to work unchanged. No functional changes.

Also added changelogs/.plugin-cache.yaml to .gitignore.
